Set up trashdb API to calculate trash score
All learning and calculating needs to be done asynchronously by default.
We can probably optimise away the async a little bit (given that we're
going to be requesting the same database content repeatedly) but by
default we should be using callbacks everywhere.

Or promises.

Yep.

Also properly implement a page to let devs debug the internal database
directly (accessible through the Chrome browser action icon in the top
right).